fix: weekly stats
